상품 삭제 API 보완

+++ src/routers/products.router.js
@@ -46,17 +46,42 @@ router.get("/products/:id", async (req, res) => {
 	})
 })
 /* 상품 수정 API */
+router.patch('/products/:id', async (req, res) => {
+	const { id } = req.params;
+	const { name, description, manager, status, password } = req.body;
+
+	const products = await Product.findById(id).exec();
+	if (products.password !== password) {
+		return res.status(404).json({ errorMessage: '비밀번호가 일치하지 않습니다.' })
+	} else {
+		if (name !== null) products.name = name;
+		if (description !== null) products.description = description;
+		if (manager !== null) products.manager = manager;
+		if (status !== null) products.status = status;
+	}
+
+	await products.save();
+
+	return res.status(200).json({ products });
+
+})
 
 /* 상품 삭제 API */
 router.delete("/products/:id", async (req, res) => {
-	const { id } = req.params;
-	const products = await Product.deleteOne({ _id: id }).exec();
+	const { id, } = req.params;
+	const { password } = req.body;
 
+	const products = await Product.findById(id).exec();
+	if (products.password !== password) {
+		return res.status(404).json({ errorMessage: '비밀번호가 일치하지 않습니다.' });
+	}
+
+	const deleteProducts = await Product.deleteOne({ _id: id }).exec();
 
 	return res.status(200).json({
 		"status": 200,
 		"message": "상품 삭제에 성공했습니다.",
-		"data": products
+		"data": deleteProducts
 	})
 })
